Added support for loading a workflow as a function in a python script.
--

Please notice that the python script must initialize VisTrails so it can execute workflows.
I've provided a load_workflow_as_function_example.py script in the scripts folder showing how to do that.
Just copy it to ./vistrails folder and change the filepaths to the vistrail file being loaded there to test it.

We use the aliases dictionary to specify the input parameters of the workflow and consequently the input parameters of the generated function:To load the workflow as a function, we can provide a function like:api.load_workflow_as_function(vtfile, workflow)     where vtfile is the path to a vistrail and workflow can be a workflow version or a workflow tagWe build the documentation of the new function based on the alias names, values and types; If there are notes available for that workflow, we can also use them in the description of the function.load_workflow_as_function is used like this:>>> isosurface = load_workflow_as_function('head.vt', 'aliases')>>> help(isosurface)Help on function isosurfaceparameters of the function:
  (diffuse_color=1.0,1.0,1.0,
   isovalue=67,
   filename=http://www.vistrails.org/download/download.php?type=DATA&id=head.120.vtk)

>>> isosurface(filename="head.vtk", isovalue="30") #this would show the spreadsheet

Eventually that should be merged with Dave's core_no_gui branch work so if the workflows don't output to the spreadsheet, the scripts won't require PyQt4 to run.

--
api/__init__.py
   New function load_workflow_as_function()