UV-CDAT: Fix for overlay plots with computed variables
This commit attempts to fix the issue that Jerry and Dean reported with plotting a computed variable (like a = clt + 2.0) with both boxfill and isoline plots.  It also contains refactoring of some of the pipeline helper code in order to minimize duplicate code.  We introduce a new connect_variables_to_plots and get_*_port_name methods in order to accomplish this.

The other major change is that variables are all created in the project_controller instead of having a mix of delayed construction of defined variables and immediate construction of computed variables.  This fixes the issue with computed variables being deleted from a pipeline on an update call (which happens with the boxfill+isoline computed variable example).

Emanuele should review these changes before merging into uvcdat.

gui/uvcdat/project_controller.py:
  - ProjectController.update_cell: create all variables (including defined variables) here; don't hardcode helper type; remove variables when not calling reset

packages/uvcdat_cdms/pipeline_helper.py:
  - CDMSPipelineHelper.get_input_port_name: new static method
  - CDMSPipelineHelper.get_plot_input_port_name: new static method
  - CDMSPipelineHelper.get_output_port_name: new static method
  - CDMSPipelineHelper.build_variable_operation_pipeline: factor operation class construction; refactor variable creation to a loop for all numbers of variables
  - CDMSPipelineHelper.connect_variables_to_plots: new method---refactored from create_actions_from_plot_objs and update_plot_pipeline_action; also changed to a nested loop for plots that take multiple variables
  - CDMSPipleineHelper.create_actions_from_plots:  only create plot_modules, connection done by connect_variables_to_plots
  - CDMSPipelineHelper.update_plot_pipeline_action: refactored connection code into connect_variables_to_plots
  - CDMSPlotWidget.__init__: set vars_were_changed to False before creating widgets