Adding support to use files from a http server
When the user write the url in the textfield and press enter, the
program will try to read the file using CDAT i.e OPenDAP protocol. For
example:
http://thredds.daac.ornl.gov/thredds/dodsC/906/totag/Totag80s.nc

However some cases files can be in a HTTP server. For example:
http://thredds.daac.ornl.gov/thredds/fileServer/906/totag/Totag80s.nc
CDAT will throw an exception in this case.

Now, UV-CDAT will check if the file is in a http server. Then, download the
file to a temporal location and open it using CDAT. The pipeline in
Vistrails connects a HTTPFile module with the CDMSVariable.

If the file was downloaded from a http server, then create HTTPFile
module and connect it to the CDMSVariable.

Check if we can download the file to a temporal location.