UV-CDAT: Fix for issue with plot overlays
This fixes an expansion of Bug #79: http://uv-cdat.llnl.gov/bugzilla/show_bug.cgi?id=79

There were two issues here.  First, the ordering that was specified by the user was not being transferred to the workflow.  Second, even with this change, a workflow does not have a deterministic ordering of inputs so even by changing the order in which plots are connected to the CDMSCell, this does not necessarily have any effect.  The first issue was corrected by using the get_plots call on the port table; the second issue required the addition of a new parameter ("plotOrder") that corresponds to the plot table ordering.  The CDMSCell now sorts its inputs based on this parameter, making sure the ordering is passed through to the lower-level plotting routines.

packages/uvcdat_cmds/init.py:
  - CDMSPlot._input_ports: add plotOrder input port
  - CDMSPlot.__init__: default plot_order to -1
  - CDMSPlot.compute: copy plotOrder parameter value to plot_order field
  - CDMSCell.compute: sort the input port list by port_order

packages/uvcdat_cdms/pipeline_helper.py:
  - CDMSPipelineHelper.create_plot_objs_from_pipeline: remove prints
  - CDMSPipelineHelper.rebuild_pipeline_action: add FIXME comment
  - CDMSPipelineHelper.update_pipeline: use user-specified ordering, persist ordering to parameters