UV-CDAT: Add mechanism for updating cell version after modifications to cell
This addresses Bug #150: http://uv-cdat.llnl.gov/bugzilla/show_bug.cgi?id=150

This code enables a cell_was_changed callback to be initiated from inside a uvcdat package.  The developer adds a call to self.controller.uvcdat_controller.cell_was_changed, passing the last action that was added.  The developer should wrap this call with a hasattr(self.controller, 'uvcdat_controller') if they wish the call to work outside of UV-CDAT.  The code is a kludge as it places a reference back to the project controller inside the vistrails controller and then takes the currently selected cell as the one to apply changes to.

gui/uvcdat/project_controller.py:
  - ProjectController.__init__: set self on vistrails controller, default current sheet name
  - ProjectController.current_cell_changed: store sheet name
  - ProjectController.cell_was_changed: new method, callback for modules in packages that wish to alert the uvcdat controller that they updated a cell

packages/vtDV3D/vtUtilities.py:
  - change_parameters: add call to cell_was_changed