<bugfix>UVCDAT Bug 183 double click variable properties</bugfix>
Double clicking on variables was no longer bringing up the variable
properties dialog.

This was due to the dialog's recent switch to being a separate window
as opposed to an undocked dockable widget. Just needed to remove a
call that used to ensure the dockable widget was floating.

gui/uvcdat/definedVariableWidget.py:
 -QDefinedVariableWidget.setupDimsForEditMode: removed call to
   self.setFloating(true) as this object is no longer a
   QDockWdiget.