<bugfix>UVCDAT Bug 194 add run duration flag</bugfix>
Aashish requested a feature that would allow users to pass
a time parameter on the command line that would quit
the application after the specified amount of time.

Added parameter -T and --time, and used a QTimer singleShot
to close the app.

vistrails/core/application.py:
 -VistrailsApplicationInterface.setupOptions: add commmand
   line option -T and --time
 -VistrailsApplicationInterface.readOptions: check option
   -T and --time and set associated variable in config

vistrails/gui/application.py:
 -VistrailsApplicationSingleton.init: if time config variable
   is greater than 0, set up a QTimer singleShot to signal
   the uvcdat main window quit slot.