Fix issues with update_vistrails call for pv contour.
The main gotcha here is the dissociation between cell versions and the controller version.  The controller is project-wide, meaning that all cells use the same controller but obviously each corresponds to a different workflow in the version tree.  The show_configuration_widget call passes the version that corresponds to the current cell, but it is currently up to the package developer to make sure that the controller is set to that version when making any calls.  Thus, you need to call controller.change_selected_version(<the version sent in from show_configuration_widget>) before making any changes to the cell.  Then, you need to notify the uvcdat controller that a change has been made via the cell_was_changed method.