<bugfix> UVCDAT Issue 76 variable edit refresh</bugfix>
When editing a variable, there was no way to apply the edits
to the variable, or to existing plots that use the variable.

Added an apply button to the edit variable widget, which updates
the variable and all plots that are using the variable.

vistrails/gui/uvcdat/definedVariableWidget.py:
-enable apply button

vistrails/gui/uvcdat/project_controller.py
-ProjectController.variableEdited: new function updates all cells
 that contain the edited variable

vistrails/gui/uvcdat/variable.py:
-VariableProperties.applyEditsClicked: new function applies edits
 to variable and calls variableEdited on controller