<bugfix>Fixed bug on PROV exporter</bugfix>
PROV exporter is not being able to look into the pipeline inside an Abstraction, and there was an error when trying to associate a module execution with a module inside an Abstraction. For now, the association is not done, and the module execution (PROV activity) is outputted even without association.

vistrails/db/services/prov.py
- create_prov: handling abstractions in get_modules_and_conn and ignoring associations in get_execs when necessary