Adds default plot feature.
Users now have the option to set a default plot.

*The setting is under UVCDAT->Preferences->Var (tab)->Default Plot
The default value is None, in which case UVCDAT behaves as normal.
*If a default plot is chosen, that plot is automatically used upon
dragging any variable into an empty cell.
*Dragging another plot into an unmodified cell constructed with a
default plot will replace the plot instead adding it to the queue.
*Any change to a cell contructed with a default plot removes it's
default 'status', in which case newly dragged plots will behave
the queue rules. (e.g. var is dropped, default plot is used,
colormap is changed, cell is no longer in a default state, dragging
another plot adds it to the queue)