GIT ERROR: previous commit was: Fixes variable deletion
Even though a messagebox was displayed to explain that the variable
couldn't be deleted, it was still removed from the UI, because the
result was not sent back to the parent.

This is the correct author/message info for the previous commit that
got a copy of a previous message by mistake.