Fix calculator to allow normal python commands
This allows the calculator to function as a python prompt for any
command that *does not* involve a defined variable.  For any defined
variable, the command is routed through the provenance framework.
Thus:

> clt2 = clt + 43

is captured as provenance, but

> e = [1,2,3]

is not.  Note, however, that doing something like loading a variable
using cdms2 outside of the variable panel is *strongly discouraged*
because it will *not* be added to the variable panel (such behavior
would require an involved rewrite of the scripting
framework). Finally, no libraries are automatically imported so if you
want to do something with cdms2, you need to *import* it first.